task_8_2/requirements.txt
aiohttp==2.0.7
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
async-timeout==1.2.1
billiard==3.3.0.23
celery==3.1.25
chardet==3.0.2
Django==1.9
django-celery==3.2.1
djangorestframework==3.6.2
kombu==3.0.37
lxml==3.7.3
multidict==2.1.4
numpy==1.12.1
packaging==16.8
pandas==0.20.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
redis==2.10.3
six==1.10.0
vine==1.1.3
yarl==0.10.1
